+++ NEWS.md
@@ -1,5 +1,7 @@
 # howManyImputations 0.2.4.9001
 
+- No need to be so strict on name; `howManyImputations()` and
+  `how_many_imputations()` are aliases.
 - Switch from **testthat** to
   [**tinytest**](https://CRAN.R-project.org/package=tinytest)